third_party/WebKit/LayoutTests/bluetooth/requestDevice/canonicalizeFilter/max-length-namePrefix-unicode.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script>
'use strict';
promise_test(() => {
  // \u00A1's UTF-8 respresentation is 2 bytes long.
  // 124 chars * 2 bytes/char = 248 bytes
  const DEVICE_NAME = '\u00A1'.repeat(124);

  return setUpPreconnectedDevice({name: DEVICE_NAME})
    .then(() => requestDeviceWithKeyDown(
        { filters: [{namePrefix: DEVICE_NAME}]}))
    .then(device => assert_equals(device.name, DEVICE_NAME));
}, 'A unicode device namePrefix of 248 bytes is valid.');
</script>
